requirements.txt
selenium>=4.0.0
python-dotenv>=0.21.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.0
webdriver-manager>=3.8.0
dotenv~=0.9.9
protobuf~=5.29.5